Makes all geometry functions and computations use floats.
Floats are faster and generally better supported on graphics hardware.
We will probably want to eventually do these computations on the gpu,
and our object data and representations will be floats... so we should
get used to it.

+++ include/apsis/geometry/circle.h
@@ -10,7 +10,7 @@ namespace Apsis {
     struct Circle {
       public:
         Point  position;
-        double radius;
+        float radius;
 
         bool intersects(Circle* circle);
